requirements.txt
streamlit==1.32.0
openai>=1.12.0
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1
python-dateutil==2.8.2 